+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Version 0.3.0 - ?
 - Create backup of data file if already exists
 - Add instant flushing of data file
+- Added instruction, how to create simple systemd service
+- Maximum file limit can be set(default 100MB), to avoid out of space problems
 
 ## Version 0.2.0 - 09.07.2023
 - Added CLI